Rename API dir to include
