+++ src/Acceleration Structures/DBVH.h
@@ -8,7 +8,7 @@
 #include <list>
 #include <utility>
 #include <deque>
-#include <API/Object.h>
+#include <include/Object.h>
 #include "../Utils/HashMap/robin_map.h"
 
 class DBVH : public Object {
