+++ src/Object/TriangleMeshObject.cpp
@@ -2,7 +2,7 @@
 // Created by sebastian on 13.11.19.
 //
 
-#include "../../API/TriangleMeshObject.h"
+#include "../../include/TriangleMeshObject.h"
 #include "../Acceleration Structures/DBVH.h"
 
 class Triangle : public Object {
